@@ -15,7 +15,7 @@
     "react-dom": "^16.13.1",
     "react-router-dom": "^6.3.0",
     "react-scripts": "3.4.1",
-    "typescript": "~3.7.2"
+    "typescript": "^4.7.2"
   },
   "homepage": "https://boevoivertolet.github.io/homework",
   "scripts": {
